Changed generation scheme for csv files with app running data.
Previously, the samples of running time for each run of an application were
recorded in different columns. Instead we now use a column for the number of the
sample, and a column for the value of the sample. This requires copying data
over several lines, one for each sample of the application.