Fixed index generation for Maps
Index generation for maps generated uniform random values on a massive range
that meant that Map applications would almost never call get() and get a result.
We generate index on twice the size of the map, so some get() calls find a
result, and some do not, which is slightly more realistic.